# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API wrapper/binding for imageboard APIs (Danbooru, Gelbooru, Rule34).",
        "title": "Matoi API Wrapper",
        "contact": {}
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "System info and health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.HomeResponse"
                        }
                    }
                }
            }
        },
        "/api/rule34/media": {
            "get": {
                "tags": [
                    "rule34"
                ],
                "summary": "Proxy and stream Rule34 media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encoded Rule34 media URL",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Streams the media file"
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Failed to fetch media",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/rule34/posts": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rule34"
                ],
                "summary": "Get posts from Rule34",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Space-separated tags",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Rule34Response"
                        }
                    },
                    "502": {
                        "description": "Upstream fetch failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Ping the server",
                "responses": {
                    "200": {
                        "description": "Server is running",
                        "schema": {
                            "$ref": "#/definitions/handlers.PingResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.HomeResponse": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "heap": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "rss": {
                    "type": "string"
                },
                "server": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "uptime": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "handlers.PingResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "handlers.Rule34Response": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Post"
                    }
                },
                "provider": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.Post": {
            "type": "object",
            "properties": {
                "directory": {
                    "type": "integer"
                },
                "file_url": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "matoi_file_url": {
                    "type": "string"
                },
                "matoi_preview_url": {
                    "type": "string"
                },
                "matoi_sample_url": {
                    "type": "string"
                },
                "preview_url": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "rating": {
                    "description": "s (safe), q (questionable), e (explicit)",
                    "type": "string"
                },
                "sample_url": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "source": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Provide the API key via the `api_key` query parameter or `Authorization: Bearer \u003ckey\u003e` header.",
            "type": "apiKey",
            "name": "api_key",
            "in": "query"
        }
    }
}